style: move functions to private.go
